+++ packages/ui/atoms/index.ts
@@ -2,6 +2,7 @@ export * from './Action/index.js';
 export * from './AnchorScrollTo/index.js';
 export * from './CircularMarquee/index.js';
 export * from './Cursor/index.js';
+export * from './Data/index.js';
 export * from './Figure/index.js';
 export * from './FigureVideo/index.js';
 export * from './LargeText/index.js';